ref(util): changed GetHTTPJSONResult implementation [#40]
GetHTTPJSONResult function was returning the data
in a shadowed manner, which defies the idiomatic ideal of golang.
So it was changed to, internally, decode the json http result as a generic
map[string]any entity and return it, delegating the responsability of
typing that piece of data to the code which uses the GetHTTPJSONResult helper.
that way we achieved a much more clean/clear/straightforward implementation